refactor:docker 서브모듈에 db 마이그레이션 파일 추가 feat: 외부 api 모듈 분리
