+++ devbie-app-api/build.gradle
@@ -2,12 +2,14 @@ dependencies {
     implementation project(':devbie-common')
     implementation project(':devbie-core')
     implementation project(':devbie-core-web')
+    implementation project(':devbie-aws-client')
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
-
+    implementation 'org.springframework.boot:spring-boot-starter-cache'
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
     compile group: 'io.springfox', name: 'springfox-swagger2', version: '2.9.2'
     compile group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.9.2'
+    compile "org.flywaydb:flyway-core"
 
     // todo 고민되는 것
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
