+++ devbie-aws-client/build.gradle
@@ -1,0 +1,9 @@
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    implementation platform('com.amazonaws:aws-java-sdk-bom:1.11.820')
+    implementation 'com.amazonaws:aws-java-sdk-s3'
+}
