+++ devbie-core/build.gradle
@@ -14,10 +14,7 @@ dependencies {
 
     implementation('org.springframework.boot:spring-boot-starter-validation')
 
-    implementation platform('com.amazonaws:aws-java-sdk-bom:1.11.820')
-    implementation 'com.amazonaws:aws-java-sdk-s3'
-
-
+    
     implementation 'org.springframework.boot:spring-boot-starter-webflux'
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
 }
